Working on rides#show but fixed other bugs as well
It turns out that there is a pretty serious bug in rails 4.1.1 when
paired with ruby 2.2. There is an issue with date parsing. Hence, I
needed to update to rails 4.2 in order to solve the issue. However, that
cause breaking changes because there was an update to Rack in the
meantime. Hence, I needed to overwrite the Rails default options for the
:Host in order to allow vagrant and the local machine to communicate and
pass data to port 3000 on the local machine.